Stop the settings patcher from running several times

+++ src/js/components/ErrorHandler.ts
@@ -105,6 +105,8 @@ export class ErrorHandler {
 
 export class Patcher {
 
+    public static version: number;
+
     /**
      * Runs patch-ups on the settings to preserve backwards compatibility.  
      * All patches MUST be documented and versioned.
@@ -113,25 +115,31 @@ export class Patcher {
 
         let counter = 0;
 
-        // Version 1.3.5
+        Patcher.version = await XM.Storage.getValue("re621.patchVersion", 0);
+
+        // Patch 1 - Version 1.3.5
         // The subscription modules were renamed to make the overall structure more clear.
         // Cache was removed from the module settings to prevent event listeners from being
         // triggered needlessly.
-        for (const type of ["Comment", "Forum", "Pool", "Tag"]) {
-            const entry = await XM.Storage.getValue("re621." + type + "Subscriptions", undefined);
-            if (entry === undefined) continue;
-            if (entry["cache"] !== undefined) {
-                await XM.Storage.setValue("re621." + type + "Tracker.cache", entry["cache"]);
-                delete entry["cache"];
+        if (Patcher.version < 1) {
+            for (const type of ["Comment", "Forum", "Pool", "Tag"]) {
+                const entry = await XM.Storage.getValue("re621." + type + "Subscriptions", undefined);
+                if (entry === undefined) continue;
+                if (entry["cache"] !== undefined) {
+                    await XM.Storage.setValue("re621." + type + "Tracker.cache", entry["cache"]);
+                    delete entry["cache"];
+                    counter++;
+                }
+                await XM.Storage.setValue("re621." + type + "Tracker", entry);
+                await XM.Storage.deleteValue("re621." + type + "Subscriptions");
                 counter++;
             }
-            await XM.Storage.setValue("re621." + type + "Tracker", entry);
-            await XM.Storage.deleteValue("re621." + type + "Subscriptions");
-            counter++;
+            Patcher.version = 1;
         }
 
         Debug.log(`Patcher: ${counter} records changed`)
 
+        await XM.Storage.setValue("re621.patchVersion", Patcher.version);
     }
 
 }
